src/requirements.txt
# API framework
fastapi>=0.95.0
uvicorn[standard]>=0.22.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.27.0
alembic>=1.10.0

# Data validation
pydantic>=2.0.0

# PDF processing
PyMuPDF>=1.21.0

# Authentication
PyJWT>=2.6.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Utility
python-dotenv>=1.0.0 